Improved the handling of vcfeval under circumstances when the number of currently unresolved paths exceeds our too-hard threshold. Previous behaviour was to bail out, and the user had to manually remove variants in the (imprecisely described) region.
We now keep better track of the bounds of the currently unresolved region,
as well as the best path up to the previous sync point. When we hit the
too-hard threshold, we revert to the best path up to the previous sync point,
and directly advance the path to the end of the region. Baseline and
calls within the region will not be represented in the output files.