Improve handling for new VA alleles found during complex call splitting.
Very rarely, during complex call splitting, a new VA is discovered which
is not present in the alleles used during initial GL map calculation. In
this case, log and skip the GL field output. This will let us more
easily find example cases. A better solution would be to recompute the
GL entirely, but now is not the time.