sammerge: allow detecting and removing duplicates via --remove-duplicates
It turns out that duplicate removal is kind of ugly. It necessarily
comes after all files are merged, and at this stage we are dealing with
ReaderRecord implementations. So we make a thin wrapper around SAMRecord
that works mostly the same way as VariantAlignmentRecord. In addition,
for backwards compatibility we need ensure that multi-file merging uses
the same ordering for SamReaderRecord iteration as regular SAMRecord
iteration.