segment: Data normalization for case and control are carried out independently.
When rows are dropped due to case or control minimum coverage
thresholds, this no longer affects the calculation of the other
sample. This ensures that, for example, control column values do not
change across different case sample runs.

The column order in the unsegmented BED file has changed as a result.